Updates benches with masp native vp
